requirements-test.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
pytest>=6.0.0
pytest-asyncio>=0.15.0
httpx>=0.19.0
python-dotenv>=0.19.0
pytest-cov>=2.12.0
